==> cpu/notebooks/1-Steady-state_design.ipynb <==
## Design a product

The design of a product is the process of defining characteristics to meet others.

Typically, geometries will be computed to reach physical specifications.


### Definition of the specs

Exchanger surface is evaluated so that maximum CPU temparature is 80°C when air temperature is 40° and CPU is running at its maximum.

from cpu.systems import CPUSystem

# create a system model
cpu = CPUSystem("cpu")
cpu

from cosapp.drivers import NonLinearSolver, RunSingleCase

# add a design driver and its definition
design = cpu.add_driver(NonLinearSolver("solver"))
runner = design.add_driver(RunSingleCase("runner"))
design.extend(cpu.design_methods["exchanger_surface"])
design.runner.set_values({"fan.T_air": 40.0, "T_cpu": 80.0, "cpu.usage": 100.0})

### The design itself

You can then run one or more designs by changing the specifications.

See also the [DoE (Design of Experiments)](doe.ipynb) to automatically run a set of cases.

# run design
cpu.run_drivers()
print(cpu.exchanger.surface)

design.runner.set_values({"fan.T_air": 30.0, "T_cpu": 80.0, "cpu.usage": 100.0})
cpu.run_drivers()
print(cpu.exchanger.surface)

from cosapp.utils.json import EncodingMetadata
from pathlib import Path

Path("data").mkdir(exist_ok=True)
cpu_json = cpu.save("data/cpu_ref.json", encoding_metadata=EncodingMetadata(with_drivers=False))
