==> cpu/notebooks/12-DoE.ipynb <==
# Design of Experiments

A Design of Experiments (DoE) is a convenient way to run multiple simulations.

## Linear DoE

The linear DoE generates an hypercube of the inputs, using a linspace.

Here we illustrate how it works with a design DoE.

from cosapp.drivers import NonLinearSolver, LinearDoE, RunSingleCase
from cosapp.recorders import DataFrameRecorder

from cpu.systems import CPUSystem

cpu = CPUSystem("cpu")

doe = cpu.add_driver(LinearDoE("doe"))

design = doe.add_child(NonLinearSolver("solver"))
runner = design.add_driver(RunSingleCase("runner"))
design.extend(cpu.design_methods["exchanger_surface"])
runner.set_values({"T_cpu": 80.0, "cpu.usage": 100.0})

doe.add_recorder(DataFrameRecorder(includes=["*"]))
doe.add_input_var({"fan.T_air": {"lower": 30.0, "upper": 60.0, "count": 31}})

cpu.run_drivers()
doe.recorder.data
